restore time remaining feature
restores a feature that was erased by the transaction list refactor,
and greatly improves upon the UX. This new version is viewable in
both fullscreen and popup mode. The component was refactored into a
hook and is consumed in the transaction-list-item component now.

Note: i'm not sure where to actually activate this feature and i'm open
to adding a new toggle to the settings screen for this one. Unless i'm
just missing it.